Add xlsx export, Docker/devcontainer tooling, frontend accessibility and formula help
- Add tpt-lattice-export-xlsx crate for writing spreadsheets
- Extend import-xlsx with richer parsing and tests
- Add Docker, docker-compose, devcontainer, and CONTRIBUTING docs
- Add accessible grid, keyboard handling, and formula help to frontend
- Add example templates and update README/todo

+++ crates/tpt-lattice-core/src/cell_value.rs
@@ -170,7 +170,7 @@ fn civil_from_days(z: i64) -> (i32, u32, u32) {
 fn days_from_civil(y: i32, m: u32, d: u32) -> i64 {
     let y = if m <= 2 { y - 1 } else { y } as i64;
     let era = if y >= 0 { y } else { y - 399 } / 400;
-    let yoe = (y - era * 400) as i64; // [0,399]
+    let yoe = y - era * 400; // [0,399]
     let m = m as i64;
     let doy = (153 * (if m > 2 { m - 3 } else { m + 9 }) + 2) / 5 + d as i64 - 1; // [0,365]
     let doe = yoe * 365 + yoe / 4 - yoe / 100 + doy; // [0,146096]
